P5.3: split AvailabilityCalendar into focused components
The 1,101-line calendar is now a 314-line orchestrator plus
MonthCalendar, CalendarLegend, NoteEditorPopover, DateActionMenu,
BulkActionsBar, and useLongPress/useNoteEditorState hooks. Both
popovers render through the shared Modal, gaining Escape-to-close and
body scroll-lock. onBulkSet contract, public props, and all data-*
test attributes preserved.
